renames page header to index.js
